Makes all imports relative
